Conditionally import the FFI version of getuid, so this doesn't break web builds
