+++ lib/src/getuid_stub.dart
@@ -1,0 +1,4 @@
+/// Gets the user ID of the current user.
+int getuid() {
+  throw 'Unable to determine UID on this system';
+}
